requirements.txt
gurobipy==10.0.1
pandas==2.0.0
PyYAML==6.0
